api: Add an API to retrieve application configuration
To keep the UI and backend decoupled, we need an API for the UI to
access elements of the server configuration.  The first user of this
will be to allow an HTTPS instance of the server to redirect to the HTTP
instance (for VNC connections).  In the future we can extend this to
support capabilities reporting as well.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>